refactor: Add wigetized layout utils
