+++ src/app-layout/internal.tsx
@@ -1,0 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+import { createWidgetizedAppLayout } from './widget';
+
+export const AppLayoutInternal = createWidgetizedAppLayout();
