+++ src/internal/utils/global-flags.ts
@@ -4,6 +4,7 @@ export const awsuiGlobalFlagsSymbol = Symbol.for('awsui-global-flags');
 
 interface GlobalFlags {
   removeHighContrastHeader?: boolean;
+  appLayoutWidget?: boolean;
 }
 
 export interface FlagsHolder {
