+++ src/split-panel/internal.tsx
@@ -1,0 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+import { createWidgetizedSplitPanel } from './widget';
+
+export const SplitPanelInternal = createWidgetizedSplitPanel();
